+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PixInternalTransactionReport.Log resource
 - priority and reason attributes to PixRequest resource
 - rules and debtorWorkspaceId parameters to CreditNote resource
+- Ledger resource
+- LedgerLog resource
+- LedgerTransaction resource
 
 ## [0.16.0] - 2026-06-01
 ### Added